Remove use_simd option from the public API
This should be a development-specific option.
This adds an environment variable, `AV_METRICS_DISABLE_SIMD`,
which can be set (to any value) to disable SIMD.

This is a breaking change.
